requirements.txt
watchdog==3.0.0
pyzbar==0.1.9
pillow==10.2.0
PyMuPDF==1.23.8
requests==2.31.0 